Optimized StaticTree with reduced loops and arrays.
git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@3017 e3ddb328-5bfe-0310-b762-aafcbcbd2528

+++ src/org/treetank/api/IConstants.java
@@ -68,7 +68,7 @@ public interface IConstants {
   public static final int INP_REFERENCE_COUNT_EXPONENT = 8;
 
   /** Exponent of pages per level (root level = 0, leaf level = 5). */
-  public static final long[] INP_LEVEL_PAGE_COUNT_EXPONENT =
+  public static final int[] INP_LEVEL_PAGE_COUNT_EXPONENT =
       {
           4 * INP_REFERENCE_COUNT_EXPONENT,
           3 * INP_REFERENCE_COUNT_EXPONENT,
